refactor: consolidate components and hooks
