+++ app/page.tsx
@@ -1,5 +1,5 @@
-import { Chat } from "@/app/components/chat/chat"
-import { LayoutApp } from "@/app/components/layout/layout-app"
+import { Chat } from "@/components/chat/chat"
+import { LayoutApp } from "@/components/layout/layout-app"
 import { MessagesProvider } from "@/lib/chat-store/messages/provider"
 
 export const dynamic = "force-dynamic"
