--- app/types/api.types.ts
@@ -1,56 +1,0 @@
-type Json =
-  | string
-  | number
-  | boolean
-  | null
-  | { [key: string]: Json | undefined }
-  | Json[]
-import type { Attachment } from "@ai-sdk/ui-utils"
-
-// DB removed
-export type SupabaseClientType = null
-
-export interface ContentPart {
-  type: string
-  text?: string
-  toolCallId?: string
-  toolName?: string
-  args?: Json
-  result?: Json
-  toolInvocation?: {
-    state: string
-    step: number
-    toolCallId: string
-    toolName: string
-    args?: Json
-    result?: Json
-  }
-  reasoning?: string
-  details?: Json[]
-}
-
-export interface Message {
-  role: "user" | "assistant" | "system" | "data" | "tool" | "tool-call"
-  content: string | null | ContentPart[]
-  reasoning?: string
-}
-
-export interface ChatApiParams {
-  model: string
-}
-
-export interface LogUserMessageParams {}
-
-export interface StoreAssistantMessageParams {}
-
-export interface ApiErrorResponse {
-  error: string
-  details?: string
-}
-
-export interface ApiSuccessResponse<T = unknown> {
-  success: true
-  data?: T
-}
-
-export type ApiResponse<T = unknown> = ApiSuccessResponse<T> | ApiErrorResponse
