+++ components/chat-input/file-items.tsx
