+++ components/chat/search-images.tsx
