+++ components/chat/utils.ts
