+++ components/layout/app-info/app-info-content.tsx
