+++ components/layout/settings/appearance/interaction-preferences.tsx
